+++ internal/app/history.go
@@ -39,6 +39,13 @@ type HistoryExportInput struct {
 	HistoryExportConfig
 }
 
+type HistoryRemoveInput struct {
+	CommonConfig
+	HistoryLocationConfig
+	HistoryRemoveConfig
+	RemoveList []string
+}
+
 // AliasList implements the alias listing functionality
 func (a *App) HistoryImport(ctx *provider.Context, input *HistoryImportInput) error {
 	zap.S().Infow("importing history")
@@ -114,10 +121,52 @@ func (a *App) HistoryExport(ctx *provider.Context, input *HistoryExportInput) er
 		historyExportList.Items = append(historyExportList.Items, *newEntry)
 		exportCount++
 	}
-	zap.S().Infof("Exporting %d entries", exportCount)
+	zap.S().Infof("exporting %d entries", exportCount)
 	err = writeExportFile(input.File, historyExportList)
 	if err != nil {
 		return fmt.Errorf("writing export file: %w", err)
+	}
+	return nil
+}
+
+func (a *App) HistoryRemove(ctx *provider.Context, input *HistoryRemoveInput) error {
+	zap.S().Infow("removing history")
+
+	historyList, err := a.historyStore.GetAll()
+	if err != nil {
+		return fmt.Errorf("getting history list: %w", err)
+	}
+	var entriesToRemove []*v1alpha1.HistoryEntry
+
+	switch {
+	case input.All:
+		for i := range historyList.Items {
+			entriesToRemove = append(entriesToRemove, &historyList.Items[i])
+		}
+	case input.Filter != "":
+		filterSpec := createFilter(input.Filter)
+		err = history.FilterHistory(historyList, filterSpec)
+		if err != nil {
+			return fmt.Errorf("filtering history list: %w", err)
+		}
+		for i := range historyList.Items {
+			entriesToRemove = append(entriesToRemove, &historyList.Items[i])
+		}
+	default:
+		for _, entryID := range input.RemoveList {
+			entry, err := a.historyStore.GetByID(entryID)
+			if err != nil {
+				return fmt.Errorf("getting history entry: %w", err)
+			}
+			entriesToRemove = append(entriesToRemove, entry)
+		}
+	}
+
+	zap.S().Infof("removing %d entries", len(entriesToRemove))
+
+	err = a.historyStore.Remove(entriesToRemove)
+	if err != nil {
+		return fmt.Errorf("removing history entries: %w", err)
 	}
 	return nil
 }